Phase 112 Part 1: Operator type safety - use enum classes directly
Changed FuncState operator method signatures to accept enum types:
- prefix(UnOpr op, ...) - was prefix(int op, ...)
- infix(BinOpr op, ...) - was infix(int op, ...)
- posfix(BinOpr op, ...) - was posfix(int op, ...)

Impact:
- Eliminated 6 redundant static_cast operations (enum→int→enum roundtrip)
- Improved type safety - prevents passing invalid operator values
- Self-documenting function signatures
- Zero performance impact - enums compile to same underlying type

Files changed:
- src/compiler/lparser.h (function signatures)
- src/compiler/lcode.cpp (implementations)
- src/compiler/parser.cpp (call sites - removed casts)

Performance: 4.49s avg (baseline 4.20s, variance 4.01-4.95s)
All tests passing: "final OK !!!"

+++ src/compiler/parser.cpp
@@ -879,7 +879,7 @@ BinOpr Parser::subexpr( expdesc *v, int limit) {
     int line = ls->getLineNumber();
     ls->nextToken();  /* skip operator */
     subexpr(v, UNARY_PRIORITY);
-    fs->prefix(static_cast<int>(uop), v, line);
+    fs->prefix(uop, v, line);
   }
   else simpleexp(v);
   /* expand while operators have priorities higher than 'limit' */
@@ -889,10 +889,10 @@ BinOpr Parser::subexpr( expdesc *v, int limit) {
     BinOpr nextop;
     int line = ls->getLineNumber();
     ls->nextToken();  /* skip operator */
-    fs->infix(static_cast<int>(op), v);
+    fs->infix(op, v);
     /* read sub-expression with higher priority */
     nextop = subexpr(&v2, priority[static_cast<int>(op)].right);
-    fs->posfix(static_cast<int>(op), v, &v2, line);
+    fs->posfix(op, v, &v2, line);
     op = nextop;
   }
   leavelevel(ls);
